refactor code after a lot of errors add driver settings

+++ src/test/java/org/example/RegisterUserTests.java
@@ -3,19 +3,19 @@ package org.example;
 import io.github.bonigarcia.wdm.WebDriverManager;
 import io.qameta.allure.junit4.DisplayName;
 import io.restassured.RestAssured;
+import io.restassured.response.Response;
 import org.example.UserInfo.UserInfo;
 import org.example.UserInfo.UserRandomizer;
 import org.example.UserInfo.UserSteps;
 import org.example.pageObject.Constants;
 import org.example.pageObject.RegistrationPage;
 import org.example.pageObject.SignInPage;
-import org.junit.After;
-import org.junit.AfterClass;
-import org.junit.Before;
-import org.junit.Test;
+import org.example.webDriverSettings.WebDriverSettings;
+import org.junit.*;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.chrome.ChromeDriver;
 
+import java.util.Random;
 import java.util.concurrent.TimeUnit;
 
 import static org.junit.Assert.assertEquals;
@@ -26,47 +26,46 @@ public class RegisterUserTests {
     public RegistrationPage registrationPage;
     public static String accessToken;
     public SignInPage signInPage;
-    public UserInfo userInfo;
+
 
     @Before
     public void setUp(){
-        WebDriverManager.chromedriver().setup();
-        driver = new ChromeDriver();
-        driver.manage().timeouts().implicitlyWait(2, TimeUnit.SECONDS);
-        registrationPage = new RegistrationPage(driver);
-        signInPage = new SignInPage(driver);
         RestAssured.baseURI = Constants.BURGER_MAIN_PAGE;
-        userInfo = UserRandomizer.userWithRandomData();
-        accessToken = UserSteps.createNewUser(userInfo).path("accessToken");
+        driver = WebDriverSettings.setBrowser();
+        registrationPage = new RegistrationPage(driver);
+        driver.get(Constants.BURGER_MAIN_PAGE+Constants.REGISTER_PAGE);
+        registrationPage.waitingForSignUpPageLoading();
     }
 
     @After
     public void teardown(){
         driver.quit();
     }
+
     @AfterClass
     public static void cleanData(){
-        UserSteps.deleteUser(accessToken);
+        if(accessToken != null){UserSteps.deleteUser(accessToken);}
+
     }
+
 
     @Test
     @DisplayName("Проверка успешной регистрации")
     public void testSuccessRegister(){
-        driver.get(Constants.BURGER_MAIN_PAGE+Constants.REGISTER_PAGE);
-        userInfo.setPassword("123456");
-        userInfo.setEmail("sdqqqqfqfq@re.uy");
-        userInfo.setName("Chak");
+        UserInfo userInfo = UserRandomizer.userWithRandomData();
+        signInPage = new SignInPage(driver);
         registrationPage.insertUserDataAndClick(userInfo);
         signInPage.waitingLoginPage();
-        assertEquals("Переход на страницу логина", Constants.BURGER_MAIN_PAGE+Constants.LOGIN_PAGE,driver.getCurrentUrl());
+        Assert.assertEquals("Перешли на страницу логина", Constants.BURGER_MAIN_PAGE + Constants.LOGIN_PAGE, driver.getCurrentUrl());
+        accessToken = UserSteps.loginUser(userInfo).path("accessToken");
     }
 
 
     @Test
     @DisplayName("Проверка вывода напдиси о некорректном пароле")
     public void testErrorAfterWrongData(){
+        UserInfo userInfo = UserRandomizer.userWithRandomData();
         userInfo.setPassword("12345");
-        driver.get(Constants.BURGER_MAIN_PAGE+Constants.REGISTER_PAGE);
         registrationPage.insertUserDataAndClick(userInfo);
         assertTrue("Ошибка о некорректном пароле отображается", registrationPage.checkWrongPasswordMessage());
         assertEquals("Пользователь остался на странице с регистрацией", Constants.BURGER_MAIN_PAGE+Constants.REGISTER_PAGE,driver.getCurrentUrl());
